+++ eslint.config.mjs
@@ -4,14 +4,7 @@ import tseslint from 'typescript-eslint'
 
 export default defineConfig(
   {
-    ignores: [
-      'dist',
-      '.nyc_output',
-      'node_modules',
-      'resources',
-      'test',
-      'test_resources',
-    ],
+    ignores: ['dist', '.nyc_output', 'node_modules', 'resources', 'test', 'test_resources'],
   },
   eslint.configs.recommended,
   tseslint.configs.recommended,